+++ app/src/main/java/woowacourse/shopping/application/ShoppingApplication.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.application
 
 import android.app.Application
-import woowacourse.shopping.data.ShoppingDatabase
-import woowacourse.shopping.data.cart.repository.CartRepositoryImpl
-import woowacourse.shopping.data.history.repository.HistoryRepositoryImpl
+import woowacourse.shopping.data.local.ShoppingDatabase
+import woowacourse.shopping.data.local.cart.repository.CartRepositoryImpl
+import woowacourse.shopping.data.local.history.repository.HistoryRepositoryImpl
 import woowacourse.shopping.feature.cart.CartViewModel
 import woowacourse.shopping.feature.goods.GoodsViewModel
 import woowacourse.shopping.feature.goodsdetails.GoodsDetailsViewModel
