+++ app/src/main/java/woowacourse/shopping/data/local/ShoppingDatabase.kt
@@ -1,13 +1,13 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.local
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import woowacourse.shopping.data.cart.CartDao
-import woowacourse.shopping.data.cart.CartEntity
-import woowacourse.shopping.data.history.HistoryDao
-import woowacourse.shopping.data.history.HistoryEntity
+import woowacourse.shopping.data.local.cart.CartDao
+import woowacourse.shopping.data.local.cart.CartEntity
+import woowacourse.shopping.data.local.history.HistoryDao
+import woowacourse.shopping.data.local.history.HistoryEntity
 
 @Database(entities = [CartEntity::class, HistoryEntity::class], version = 1)
 abstract class ShoppingDatabase : RoomDatabase() {
