+++ app/src/main/java/woowacourse/shopping/data/local/cart/CartDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping.data.local.cart
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
