+++ app/src/main/java/woowacourse/shopping/data/local/cart/CartEntity.kt
@@ -1,15 +1,14 @@
-package woowacourse.shopping.data.history
+package woowacourse.shopping.data.local.cart
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "history")
-data class HistoryEntity(
+@Entity(tableName = "cart")
+data class CartEntity(
     @PrimaryKey val id: Long,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "price") val price: Int,
     @ColumnInfo(name = "thumbnail_url") val thumbnailUrl: String,
     @ColumnInfo(name = "quantity") val quantity: Int,
-    @ColumnInfo(name = "timestamp") val timestamp: Long = System.currentTimeMillis(),
 )
