+++ app/src/main/java/woowacourse/shopping/data/local/history/HistoryDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.history
+package woowacourse.shopping.data.local.history
 
 import androidx.room.Dao
 import androidx.room.Insert
