+++ app/src/main/java/woowacourse/shopping/data/local/history/HistoryEntity.kt
@@ -1,14 +1,15 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping.data.local.history
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "cart")
-data class CartEntity(
+@Entity(tableName = "history")
+data class HistoryEntity(
     @PrimaryKey val id: Long,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "price") val price: Int,
     @ColumnInfo(name = "thumbnail_url") val thumbnailUrl: String,
     @ColumnInfo(name = "quantity") val quantity: Int,
+    @ColumnInfo(name = "timestamp") val timestamp: Long = System.currentTimeMillis(),
 )
