+++ app/src/main/java/woowacourse/shopping/data/local/history/repository/HistoryRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.history.repository
+package woowacourse.shopping.data.local.history.repository
 
 import woowacourse.shopping.domain.model.Cart
