+++ app/src/main/java/woowacourse/shopping/feature/cart/CartViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.switchMap
-import woowacourse.shopping.data.cart.repository.CartRepository
+import woowacourse.shopping.data.local.cart.repository.CartRepository
 import woowacourse.shopping.domain.model.Cart
 import woowacourse.shopping.domain.model.Carts
 import woowacourse.shopping.util.updateQuantity
