+++ app/src/main/java/woowacourse/shopping/feature/goodsdetails/GoodsDetailsViewModel.kt
@@ -3,8 +3,8 @@ package woowacourse.shopping.feature.goodsdetails
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.cart.repository.CartRepository
-import woowacourse.shopping.data.history.repository.HistoryRepository
+import woowacourse.shopping.data.local.cart.repository.CartRepository
+import woowacourse.shopping.data.local.history.repository.HistoryRepository
 import woowacourse.shopping.domain.model.Cart
 import woowacourse.shopping.domain.model.Goods
 import woowacourse.shopping.util.MutableSingleLiveData
